filtros

+++ blog/apps/noticias/views.py
@@ -8,6 +8,7 @@ from .forms import ContactoForm, NoticiaForm
 # importamos reverse lazy para los comentarios
 from django.urls import reverse_lazy
 
+from django.db.models import Q 
 
 # decorador para ver las noticias solamente como usuario logueado
 from django.contrib.auth.decorators import login_required
@@ -16,31 +17,54 @@ from django.views import View
 
 
 # uso de decorador para verificar logeo de usuario y poder ver noticia
-@login_required
+# @login_required
+# def inicio(request):
+#     # obtener todas las noticias y mostrar en el inicio.html
+#     # ctx = {}
+#     # # clase.objetcs.all()==> select * from noticia
+#     # noticia = Noticia.objects.all()
+#     # ctx["noticias"] = noticia
+#     # return render(request, 'noticias/inicio.html', ctx)
+#     contexto = {}
+#     id_categoria = request.GET.get('id', None)
+
+#     if id_categoria:
+#         n = Noticia.objects.filter(categoria_noticia=id_categoria)
+#     else:
+#         n = Noticia.objects.all().order_by('-fecha')  # una lista
+
+#     contexto['noticias'] = n
+    
+#     cat = Categoria.objects.all().order_by('nombre')
+#     contexto['categorias'] = cat
+
+#     return render(request, 'noticias/inicio.html', contexto)
 def inicio(request):
-    # obtener todas las noticias y mostrar en el inicio.html
-    # ctx = {}
-    # # clase.objetcs.all()==> select * from noticia
-    # noticia = Noticia.objects.all()
-    # ctx["noticias"] = noticia
-    # return render(request, 'noticias/inicio.html', ctx)
     contexto = {}
-    id_categoria = request.GET.get('id', None)
+    queryset = request.GET.get("buscar") 
+    if queryset:
+        n = Noticia.objects.filter(
+            Q(titulo__icontains = queryset) |
+            Q(cuerpo__icontains = queryset) |
+            Q(categoria_noticia__nombre = queryset)
+        ).distinct().order_by('-fecha')[:5]
+        contexto['noticias'] = n
 
-    if id_categoria:
-        n = Noticia.objects.filter(categoria_noticia=id_categoria)
+        return render(request, 'noticias/inicio.html',contexto)
     else:
-        n = Noticia.objects.all()  # una lista
+        id_categoria = request.GET.get('id', None)
+        if id_categoria:
+            n = Noticia.objects.filter(categoria_noticia=id_categoria).order_by('-fecha')[:5]
+        else:
+            n = Noticia.objects.all().order_by('-fecha')[:5]  # una lista
+        contexto['noticias'] = n
+        cat = Categoria.objects.all().order_by('nombre')[:5]
+        contexto['categorias'] = cat
+        return render(request, 'noticias/inicio.html', contexto)
 
-    contexto['noticias'] = n
-
-    cat = Categoria.objects.all().order_by('nombre')
-    contexto['categorias'] = cat
-
-    return render(request, 'noticias/inicio.html', contexto)
 
 
-@login_required
+# @login_required
 def Detalle_Noticias(request, pk):
     contexto = {}
 
@@ -85,7 +109,8 @@ def Comentar_Noticia(request):
 class Editcomentario(View):
     def get(self, request, pk):
         comment = Comentario.objects.get(pk=pk) # extraemos el objeto de comentarios con igual pk
-        return render(request, 'noticias/edit.html', {'comment': comment})
+        noticia = comment.noticia 
+        return render(request, 'noticias/edit.html', {'comment': comment, 'noticia': noticia})
 
     def post(self, request, pk):
         comment = Comentario.objects.get(pk=pk)
@@ -100,8 +125,9 @@ class Editcomentario(View):
 class CommentDeleteView(View):
     def get(self, request, pk):
         comment = get_object_or_404(Comentario, pk=pk) # comprobamos si el usuaro que va a borrarlo es el mismo que quien creó el comentario, para mandarlo al html de delete
+        noticia = comment.noticia
         if comment.usuario == request.user:
-            return render(request, 'noticias/delete.html', {'comment': comment})
+            return render(request, 'noticias/delete.html', {'comment': comment, 'noticia': noticia})
         else:
             noticia = comment.noticia  # sinó lo redireccionamos de nuevo
             return redirect('noticias:detalle', pk=noticia.pk) 
@@ -109,11 +135,12 @@ class CommentDeleteView(View):
     def post(self, request, pk):
         comment = get_object_or_404(Comentario, pk=pk)
         if comment.usuario == request.user: # si el usuario es el que creó el comentario este se borra
-            comment.delete()
+            if request.method == 'POST':
+                comment.delete()
         noticia = comment.noticia
         return redirect('noticias:detalle', pk=noticia.pk)  
 
-
+# agregar, eliminar, editar y filtrar noticias
 def agregar_noticia(request):
     if request.method == 'POST':
         form = NoticiaForm(request.POST, request.FILES)
@@ -130,9 +157,50 @@ def agregar_noticia(request):
 
 def eliminar_noticia(request, pk):
     noticia = Noticia.objects.get(pk =pk)
-    noticia.delete()
-    return redirect("noticias:inicio")
+    noticia = get_object_or_404(Noticia, pk=pk)
+    contexto = {'noticia': noticia}
+    if request.method == 'POST':
+        noticia.delete()
+        return redirect('noticias:inicio')
+    return render(request, 'noticias/deletenoticia.html', contexto)
+
+ 
 
 
+def editar_noticia(request, pk):
+    # obtengo la noticia y si no existe tira error
+    noticia = get_object_or_404(Noticia, pk=pk)
+    # si se aprieta el boton de guardar cambios manda la info a travez de POST 
+    if request.method == 'POST':
+        # inicializamos el formulario con el parametro instance, que es un parametro que se necesita para actualizar una instancia de la BD
+        form = NoticiaForm(request.POST, instance=noticia)
+        if form.is_valid():
+            form.save()
+            return redirect('noticias:detalle', pk=noticia.pk) 
+
+    else:
+        form = NoticiaForm(instance=noticia)
+   
+    return render(request, 'noticias/editnoticia.html',{'form': form, 'noticia': noticia})
 
 
+def filtrar_noticia(request):
+    fecha = request.GET.get('fecha')
+    orden = request.GET.get('orden')
+
+    if fecha == 'asc':
+        noticias = Noticia.objects.all().order_by('fecha')
+     
+    elif fecha == 'des':
+        noticias= Noticia.objects.all().order_by('-fecha')
+     
+
+    if orden == 'asc':
+        noticias = Noticia.objects.all().order_by('titulo')
+    
+    elif orden == 'des':
+        noticias = Noticia.objects.all().order_by('-titulo')
+      
+
+    return render(request, 'noticias/inicio.html',{ 'noticias': noticias})
+
